--- hr-payroll/src/main/java/com/jsoftware/hrpayroll/HrPayrollApplication.java
@@ -1,15 +1,0 @@
-package com.jsoftware.hrpayroll;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.openfeign.EnableFeignClients;
-
-@EnableFeignClients
-@SpringBootApplication
-public class HrPayrollApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(HrPayrollApplication.class, args);
-	}
-
-}
