--- hr-payroll/src/main/java/com/jsoftware/hrpayroll/config/AppConfig.java
@@ -1,14 +1,0 @@
-package com.jsoftware.hrpayroll.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.client.RestTemplate;
-
-@Configuration
-public class AppConfig {
-
-	@Bean
-	public RestTemplate restTemplate() {
-		return new RestTemplate();
-	}
-}
