--- hr-payroll/src/main/java/com/jsoftware/hrpayroll/config/WorkerFeignClient.java
@@ -1,18 +1,0 @@
-package com.jsoftware.hrpayroll.config;
-
-import org.springframework.cloud.openfeign.FeignClient;
-import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Component;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-
-import com.jsoftware.hrpayroll.entities.Worker;
-
-@Component
-@FeignClient(name="hr-worker", url = "localhost:8001", path = "/workers")
-public interface WorkerFeignClient {
-
-	@GetMapping(value = "/{id}")
-	ResponseEntity<Worker> findById(@PathVariable Long id);
-	
-}
