--- hr-payroll/src/main/java/com/jsoftware/hrpayroll/entities/Payment.java
@@ -1,26 +1,0 @@
-package com.jsoftware.hrpayroll.entities;
-
-import java.io.Serializable;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class Payment implements Serializable{
-
-	/**
-	 * 
-	 */
-	private static final long serialVersionUID = 1L;
-	
-	private String name;
-	private Double dailyIncome;
-	private Integer days;
-
-	public Double getTotal() {
-		return days * dailyIncome;
-	}
-}
