--- hr-payroll/src/main/java/com/jsoftware/hrpayroll/entities/Worker.java
@@ -1,25 +1,0 @@
-package com.jsoftware.hrpayroll.entities;
-
-import java.io.Serializable;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class Worker implements Serializable{
-
-	/**
-	 * 
-	 */
-	private static final long serialVersionUID = 1L;
-
-	private Long id;
-	
-	private String name;
-	
-	private Double dailyIncome;
-	
-}
