--- hr-payroll/src/test/java/com/jsoftware/hrpayroll/HrPayrollApplicationTests.java
@@ -1,13 +1,0 @@
-package com.jsoftware.hrpayroll;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class HrPayrollApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
