--- hr-worker/src/main/java/com/jsoftware/hrworker/HrWorkerApplication.java
@@ -1,13 +1,0 @@
-package com.jsoftware.hrworker;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class HrWorkerApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(HrWorkerApplication.class, args);
-	}
-
-}
