--- hr-worker/src/main/java/com/jsoftware/hrworker/entities/Worker.java
@@ -1,35 +1,0 @@
-package com.jsoftware.hrworker.entities;
-
-import java.io.Serializable;
-
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.Table;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-@Table(name = "tb_worker")
-@Entity
-public class Worker implements Serializable{
-
-	/**
-	 * 
-	 */
-	private static final long serialVersionUID = 1L;
-
-	@Id
-	@GeneratedValue(strategy=GenerationType.IDENTITY)
-	private Long id;
-	
-	private String name;
-	
-	private Double dailyIncome;
-	
-}
