--- hr-worker/src/main/java/com/jsoftware/hrworker/repositories/WorkerRepository.java
@@ -1,9 +1,0 @@
-package com.jsoftware.hrworker.repositories;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import com.jsoftware.hrworker.entities.Worker;
-
-public interface WorkerRepository extends JpaRepository<Worker, Long>{
-
-}
