--- hr-worker/src/test/java/com/jsoftware/hrworker/HrWorkerApplicationTests.java
@@ -1,13 +1,0 @@
-package com.jsoftware.hrworker;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class HrWorkerApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
