+++ hr_payroll/src/main/java/com/jsoftware/hr_payroll/HrPayrollApplication.java
@@ -1,0 +1,15 @@
+package com.jsoftware.hr_payroll;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.openfeign.EnableFeignClients;
+
+@EnableFeignClients
+@SpringBootApplication
+public class HrPayrollApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(HrPayrollApplication.class, args);
+	}
+
+}
