+++ hr_payroll/src/main/java/com/jsoftware/hr_payroll/config/WorkerFeignClient.java
@@ -1,0 +1,18 @@
+package com.jsoftware.hr_payroll.config;
+
+import org.springframework.cloud.openfeign.FeignClient;
+import org.springframework.http.ResponseEntity;
+import org.springframework.stereotype.Component;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+
+import com.jsoftware.hr_payroll.entities.Worker;
+
+@Component
+@FeignClient(name="hr-worker", url = "localhost:8001", path = "/workers")
+public interface WorkerFeignClient {
+
+	@GetMapping(value = "/{id}")
+	ResponseEntity<Worker> findById(@PathVariable Long id);
+	
+}
