+++ hr_payroll/src/main/java/com/jsoftware/hr_payroll/entities/Payment.java
@@ -1,0 +1,26 @@
+package com.jsoftware.hr_payroll.entities;
+
+import java.io.Serializable;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class Payment implements Serializable{
+
+	/**
+	 * 
+	 */
+	private static final long serialVersionUID = 1L;
+	
+	private String name;
+	private Double dailyIncome;
+	private Integer days;
+
+	public Double getTotal() {
+		return days * dailyIncome;
+	}
+}
