+++ hr_payroll/src/main/java/com/jsoftware/hr_payroll/entities/Worker.java
@@ -1,0 +1,25 @@
+package com.jsoftware.hr_payroll.entities;
+
+import java.io.Serializable;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class Worker implements Serializable{
+
+	/**
+	 * 
+	 */
+	private static final long serialVersionUID = 1L;
+
+	private Long id;
+	
+	private String name;
+	
+	private Double dailyIncome;
+	
+}
