+++ hr_payroll/src/test/java/com/jsoftware/hr_payroll/HrPayrollApplicationTests.java
@@ -1,0 +1,13 @@
+package com.jsoftware.hr_payroll;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class HrPayrollApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
