+++ hr_worker/.gitignore
