+++ hr_worker/src/main/java/com/jsoftware/hr_worker/HrWorkerApplication.java
@@ -1,0 +1,13 @@
+package com.jsoftware.hr_worker;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class HrWorkerApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(HrWorkerApplication.class, args);
+	}
+
+}
