+++ hr_worker/src/main/java/com/jsoftware/hr_worker/resources/WorkerResource.java
@@ -1,0 +1,34 @@
+package com.jsoftware.hr_worker.resources;
+
+import java.util.List;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+import com.jsoftware.hr_worker.entities.Worker;
+import com.jsoftware.hr_worker.repositories.WorkerRepository;
+
+@RestController
+@RequestMapping(value= "/workers")
+public class WorkerResource {
+
+	@Autowired
+	private WorkerRepository repository;
+	
+	@GetMapping
+	public ResponseEntity<List<Worker>> findAll(){
+		List<Worker> workers = repository.findAll();
+		return ResponseEntity.ok().body(workers);
+	}
+	
+	@GetMapping(value = "/{id}")
+	public ResponseEntity<Worker> findById(@PathVariable Long id){
+		Worker worker = repository.findById(id).get();
+		return ResponseEntity.ok().body(worker);
+	}
+	
+}
