+++ hr_worker/src/test/java/com/jsoftware/hr_worker/HrWorkerApplicationTests.java
@@ -1,0 +1,13 @@
+package com.jsoftware.hr_worker;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class HrWorkerApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
